requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
selenium==4.7.2